--- src/main/java/pl/games/checkers/Move.java
@@ -1,25 +1,0 @@
-package pl.games.checkers;
-
-public class Move {
-
-    private final MoveType type;
-    private final Pawn killedPawn;
-
-    public Move(MoveType type) {
-        this(type, null);
-    }
-
-    public Move(MoveType type, Pawn killedPawn) {
-        this.type = type;
-        this.killedPawn = killedPawn;
-    }
-
-    public MoveType type() {
-        return type;
-    }
-
-    public Pawn killedPawn() {
-        return killedPawn;
-    }
-
-}
