--- src/main/java/pl/games/checkers/MoveType.java
@@ -1,9 +1,0 @@
-package pl.games.checkers;
-
-public enum MoveType {
-
-    INVALID,
-    MOVE,
-    KILL
-
-}
