--- src/main/java/pl/games/checkers/TileType.java
@@ -1,16 +1,0 @@
-package pl.games.checkers;
-
-import javafx.scene.paint.Color;
-
-public enum TileType {
-
-    ALLOWED(Color.valueOf("#005066")),
-    DISALLOWED(Color.valueOf("#78cce2"));
-
-    final Color background;
-
-    TileType(Color background) {
-        this.background = background;
-    }
-
-}
