+++ src/main/java/pl/games/checkers/ai/PawnMoveRecursive.java
@@ -3,6 +3,7 @@ package pl.games.checkers.ai;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import pl.games.checkers.*;
+import pl.games.checkers.model.*;
 
 import java.util.*;
 import java.util.concurrent.ForkJoinTask;
