+++ src/main/java/pl/games/checkers/model/MoveType.java
@@ -1,0 +1,9 @@
+package pl.games.checkers.model;
+
+public enum MoveType {
+
+    INVALID,
+    MOVE,
+    KILL
+
+}
