+++ src/main/java/pl/games/checkers/ui/Tile.java
@@ -1,0 +1,37 @@
+package pl.games.checkers.ui;
+
+import javafx.scene.shape.Rectangle;
+import pl.games.checkers.model.TileType;
+
+public class Tile extends Rectangle {
+
+    private final TileType type;
+    private PawnFigure pawn;
+
+    public PawnFigure getPawn() {
+        return pawn;
+    }
+
+    public void setPawn(PawnFigure pawn) {
+        this.pawn = pawn;
+    }
+
+    public Tile(int row, int column) {
+        setWidth(Checkerboard.TILE_SIZE_X);
+        setHeight(Checkerboard.TILE_SIZE_Y);
+
+        relocate(column * Checkerboard.TILE_SIZE_X, row * Checkerboard.TILE_SIZE_Y);
+
+        this.type = isNotAllowed(row, column) ? TileType.DISALLOWED : TileType.ALLOWED;
+        setFill(type.getBackground());
+    }
+
+    public boolean isAllowed() {
+        return type == TileType.ALLOWED;
+    }
+
+    public static boolean isNotAllowed(int y, int x) {
+        return (y + x) % 2 == 0;
+    }
+
+}
